+++ components/menu-dropdown/menu-dropdown.jsx
@@ -64,6 +64,15 @@ overlay.style.position = 'absolute';
 
 let currentOpenDropdown;
 
+const DropdownNubbinPositions = [
+	'top left',
+	'top',
+	'top right',
+	'bottom left',
+	'bottom',
+	'bottom right'
+];
+
 /**
  * The MenuDropdown component is a variant of the Lightning Design System Menu component.
  */
@@ -171,11 +180,15 @@ const MenuDropdown = React.createClass({
 		 */
 		length: PropTypes.oneOf([null, '5', '7', '10']),
 		/**
+		 * Style applied to menu element (that is the `.slds-dropdown` element)
+		 */
+		menuStyle: PropTypes.object,
+		/**
 		 * Renders menu within an absolutely positioned container at an elevated z-index.
 		 */
 		modal: PropTypes.bool,
 		/**
-		 * Positions dropdown menu with a nubbin--that is the arrow notch. The placement options correspond to the placement of the nubbin. This is implemeted with CSS classes and is best used with a `Button` with "icon container" styling. Dropdown menus will still be contained to the closest scrolling parent.
+		 * Positions dropdown menu with a nubbin--that is the arrow notch. The placement options correspond to the placement of the nubbin. This is implemeted with CSS classes and is best used with a `Button` with "icon container" styling (`iconVariant="container"`). Use with `modal={false}`, since positioning is determined by CSS via absolute-relative positioning, and using an absolutely positioned menu will not position the menu correctly.
 		 */
 		nubbinPosition: PropTypes.oneOf([
 			'top left',
@@ -364,9 +377,10 @@ const MenuDropdown = React.createClass({
 	handleMouseEnter (event) {
 		this.isHover = true;
 
-		if (!this.state.isOpen) {
+		if (!this.state.isOpen && this.props.openOn === 'hover') {
 			this.handleOpen();
 		} else {
+			// we want this clear when openOn is hover or hybrid
 			clearTimeout(this.isClosing);
 		}
 
@@ -407,6 +421,15 @@ const MenuDropdown = React.createClass({
 
 		if (this.props.onFocus) {
 			this.props.onFocus(event);
+		}
+	},
+
+	handleClickCustomContent () {
+		this.setFocus();
+		this.handleClose();
+
+		if (this.props.onSelect) {
+			this.props.onSelect();
 		}
 	},
 
@@ -472,10 +495,12 @@ const MenuDropdown = React.createClass({
 		}
 	},
 
+	// Trigger opens, closes, and recieves focus on close
 	saveRefToTrigger (trigger) {
 		this.trigger = trigger;
 	},
 
+	// TriggerContainer is the wrapping outer DOM element which may differ from the actual trigger which is most likely a `button`.
 	saveRefToTriggerContainer (triggerContainer) {
 		this.triggerContainer = triggerContainer;
 		if (!this.trigger) this.trigger = triggerContainer;
@@ -534,7 +559,11 @@ const MenuDropdown = React.createClass({
 			if (child && child.type.displayName === LIST) {
 				customContentWithListPropInjection.push(this.renderDefaultPopoverContent(child.props));
 			} else {
-				customContentWithListPropInjection.push(child);
+				const clonedCustomContent = React.cloneElement(child, {
+					onClick: this.handleClickCustomContent,
+					key: shortid.generate()
+				});
+				customContentWithListPropInjection.push(clonedCustomContent);
 			}
 		});
 		if (customContentWithListPropInjection.length === 0) {
@@ -545,10 +574,26 @@ const MenuDropdown = React.createClass({
 	},
 
 	renderSimplePopover (customContent, isOpen) {
+		let marginTop;
+		let positionClassName;
+		if (this.props.nubbinPosition) {
+			const positions = this.props.nubbinPosition.split(' ');
+			positionClassName = classNames(
+				`slds-nubbin--${positions.join('-')}`,
+				positions.map((position) => `slds-dropdown--${position}`)
+			);
+			marginTop = 0;
+			// TODO: allow nubbinPosition prop to set the offset automatically
+			// if (this.props.nubbinPosition === 'top right') {
+			// 	offset = '-12px -24px';
+			// }
+		} else if (this.props.align) {
+			positionClassName = `slds-dropdown--${this.props.align}`;
+		}
 		return (
 			isOpen ?
 				<div
-					className={classNames('slds-dropdown', `slds-dropdown--${this.props.align}`, this.props.className)}
+					className={classNames('slds-dropdown', positionClassName, this.props.className)}
 					onMouseEnter={(this.props.openOn === 'hover') ? this.handleMouseEnter : null}
 					onMouseLeave={(this.props.openOn === 'hover') ? this.handleMouseLeave : null}
 					style={this.props.menuStyle}
@@ -597,6 +642,7 @@ const MenuDropdown = React.createClass({
 					onMouseEnter={(this.props.openOn === 'hover') ? this.handleMouseEnter : null}
 					onMouseLeave={(this.props.openOn === 'hover') ? this.handleMouseLeave : null}
 					outsideClickIgnoreClass={outsideClickIgnoreClass}
+					style={this.props.menuStyle}
 					targetElement={this.triggerContainer}
 				>
 					{this.renderPopoverContent(customContent)}
@@ -619,6 +665,7 @@ const MenuDropdown = React.createClass({
 	render () {
 		// Dropdowns are used by other components. The default trigger is a button, but some other components use `li` elements. The following allows `MenuDropdown` to be extended by providing a child component with the displayName of `DropdownTrigger`.
 		let CurrentTrigger = DefaultTrigger;
+
 		let CustomTriggerChildProps = {};
 
 		// Child elements that do not have the display name of the value of `MENU_DROPDOWN_TRIGGER` in `components/constants.js` will be considered custom content and rendered in the popover.
@@ -670,6 +717,8 @@ const MenuDropdown = React.createClass({
 				inverse={this.props.buttonInverse}
 				isOpen={isOpen}
 				label={this.props.label}
+				openOn={this.props.openOn}
+				isInline={!this.props.modal}
 				style={this.props.style}
 				tabIndex={isOpen ? '-1' : '0'}
 				variant={this.props.buttonVariant}
@@ -687,7 +736,10 @@ const MenuDropdown = React.createClass({
 				onFocus={this.props.openOn === 'hover' ? this.handleFocus : null}
 				onKeyDown={this.handleKeyDown}
 				onMouseDown={this.props.onMouseDown}
-				onMouseEnter={this.props.openOn === 'hover' ? this.handleMouseEnter : null}
+				onMouseEnter={(this.props.openOn === 'hover' || this.props.openOn === 'hybrid')
+					? this.handleMouseEnter
+					: null
+				}
 				onMouseLeave={
 					this.props.openOn === 'hover'
 					|| this.props.openOn === 'hybrid'
@@ -705,3 +757,4 @@ const MenuDropdown = React.createClass({
 module.exports = MenuDropdown;
 module.exports.ListItem = ListItem;
 module.exports.ListItemLabel = ListItemLabel;
+module.exports.DropdownNubbinPositions = DropdownNubbinPositions;