+++ stories/dropdown/index.jsx
@@ -6,6 +6,7 @@ import { storiesOf, action } from '@kadira/storybook';
 
 import { MENU_DROPDOWN } from '../../utilities/constants';
 import Dropdown from '../../components/menu-dropdown';
+import { DropdownNubbinPositions } from '../../components/menu-dropdown/menu-dropdown';
 import List from '../../components/menu-list/list';
 import Button from '../../components/button';
 import Trigger from '../../components/menu-dropdown/button-trigger';
@@ -27,76 +28,42 @@ const getDropdown = (props) => (
 	<Dropdown {...props} />
 );
 
-const DropdownControlled = React.createClass({
-	displayName: 'DropdownControlled',
-	getInitialState () {
-		return {
-			isOpen: undefined
-		};
-	},
-	handleClickOpen () {
-		this.setState({ isOpen: true });
-	},
-	handleClickClose () {
-		this.setState({ isOpen: false });
-	},
-	handleClickReset () {
-		this.setState({ isOpen: null });
-	},
-	handleOpen () {
-		this.setState({ isOpen: true });
-	},
-	handleClose () {
-		this.setState({ isOpen: undefined });
-	},
-	handleClickCustomContent () {
-		this.setState({ isOpen: false });
-	},
-	render () {
-		let isOpen;
-		if (this.state && this.state.isOpen === true) {
-			isOpen = true;
-		} else if (this.state && this.state.isOpen === false) {
-			isOpen = false;
-		}	else {
-			isOpen = undefined;
-		}
-		const props = this.props;
-		return (
-			<div className="slds-grid">
-
-				<div className="slds-col">
-					<Dropdown
-						{...props}
-						isOpen={isOpen}
-						onClose={this.handleClose}
-						onOpen={this.handleOpen}
-					>
-						<div id="custom-dropdown-menu-content">
-							<div className="slds-m-around--medium">
-								<div className="slds-tile slds-tile--board slds-m-horizontal--small">
-									<p className="tile__title slds-text-heading--small">Art Vandelay</p>
-									<div className="slds-tile__detail">
-										<p className="slds-truncate">
-											<a className="slds-m-right--medium" onClick={this.handleClickCustomContent}>Settings</a>
-											<a href="#" >Log Out</a>
-										</p>
-									</div>
-								</div>
-							</div>
-						</div>
-						<List options={options} />
-					</Dropdown>
-				</div>
-				<div className="slds-col">
-					<Button label="Open Dropdown" onClick={this.handleClickOpen} />
-					<Button label="Close Dropdown" onClick={this.handleClickClose} />
-					<Button label="Reset Dropdown" onClick={this.handleClickReset} />
-				</div>
+const getDropdownPositioned = (props) => {
+	const positionedDropdowns = [];
+	DropdownNubbinPositions.forEach((position) => {
+		positionedDropdowns.push(
+			<div className="slds-col slds-size--1-of-3" style={{ minHeight: '500px' }}>
+				<Dropdown
+					{...props}
+					forceOpen
+					nubbinPosition={position}
+				>
+					<Trigger>
+						<Button iconVariant="container" iconName="settings" label={position} />
+					</Trigger>
+				</Dropdown>
 			</div>
 		);
-	}
-});
+	});
+
+	return (
+		<div>
+			<div className="slds-grid slds-wrap">
+				{positionedDropdowns}
+			</div>
+			<div className="slds-col" style={{ minHeight: '500px' }}>
+				<Dropdown
+					{...props}
+					nubbinPosition="top right"
+				>
+					<Trigger>
+						<Button iconVariant="container" iconName="settings" assistiveText="top right" />
+					</Trigger>
+				</Dropdown>
+			</div>
+		</div>
+	);
+};
 
 const getDropdownCustomTrigger = (props) => (
 	<Dropdown {...props} >
@@ -106,21 +73,27 @@ const getDropdownCustomTrigger = (props) => (
 	</Dropdown>
 );
 
-const getDropdownCustomContent = (props) => (
-	<Dropdown {...props} >
-		<div id="custom-dropdown-menu-content">
-			<div className="slds-m-around--medium">
-				<div className="slds-tile slds-tile--board slds-m-horizontal--small">
-					<p className="tile__title slds-text-heading--small">Art Vandelay</p>
-					<div className="slds-tile__detail">
-						<p className="slds-truncate">
-							<a className="slds-m-right--medium" href="#">Settings</a>
-							<a href="#" >Log Out</a>
-						</p>
-					</div>
+/* eslint-disable react/prop-types */
+/* eslint-disable no-script-url */
+const DropdownCustomContent = (props) => (
+	<div id="custom-dropdown-menu-content">
+		<div className="slds-m-around--medium">
+			<div className="slds-tile slds-tile--board slds-m-horizontal--small">
+				<p className="tile__title slds-text-heading--small">Art Vandelay</p>
+				<div className="slds-tile__detail">
+					<p className="slds-truncate">
+						<a className="slds-m-right--medium" href="javascript:void(0)" onClick={props.onClick}>Settings</a>
+						<a href="javascript:void(0)" onClick={props.onClick}>Log Out</a>
+					</p>
 				</div>
 			</div>
 		</div>
+	</div>
+);
+
+const getDropdownCustomContent = (props) => (
+	<Dropdown {...props} >
+		<DropdownCustomContent />
 		<List options={[{ label: 'Custom Content Option' }, ...options]} />
 	</Dropdown>
 );
@@ -144,11 +117,13 @@ storiesOf(MENU_DROPDOWN, module)
 		modal: false,
 		options
 	}))
-	.add('Controled w/ isOpen', () => <DropdownControlled
-		align="right"
-		label="Dropdown Click"
-		options={options}
-	/>)
+	.add('No Modal w/ Nubbins', () => getDropdownPositioned({
+		onSelect: (...rest) => {
+			action('Selected')(...rest);
+		},
+		modal: false,
+		options
+	}))
 	.add('Custom Trigger', () => getDropdownCustomTrigger({
 		assistiveText: 'Custom Dropdown Trigger',
 		onSelect: (...rest) => {