+++ stories/global-navigation-bar/index.jsx
@@ -26,7 +26,7 @@ const dropdownCollection = [{
 	label: 'Main action', value: '0', iconCategory: 'utility',
 	iconName: 'table', href: 'http://www.google.com'
 }, {
-	label: 'Menu Header', type: 'header'
+	label: 'Menu Header', type: 'header', divider: 'top'
 }, {
 	label: 'Menu Item One', value: '1', iconCategory: 'utility',
 	iconName: 'kanban', href: 'http://www.google.com'
@@ -39,6 +39,8 @@ const dropdownCollection = [{
 }, {
 	label: 'Menu Item Four', value: '4', iconCategory: 'utility',
 	iconName: 'side_list', href: 'http://www.google.com'
+}, {
+	type: 'divider'
 }, {
 	label: 'Menu Item Five', value: '5', iconCategory: 'utility',
 	iconName: 'side_list', href: 'http://www.google.com'